+++ subscription/inmemory/src/main/java/org/occurrent/subscription/inmemory/InMemorySubscription.java
@@ -1,0 +1,69 @@
+/*
+ * Copyright 2020 Johan Haleby
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.occurrent.subscription.inmemory;
+
+import org.occurrent.subscription.api.blocking.Subscription;
+
+import java.time.Duration;
+import java.util.Objects;
+import java.util.StringJoiner;
+
+/**
+ * An in-memory subscription
+ */
+public class InMemorySubscription implements Subscription {
+    private final String id;
+
+    InMemorySubscription(String id) {
+        this.id = id;
+    }
+
+    @Override
+    public String id() {
+        return id;
+    }
+
+    @Override
+    public void waitUntilStarted() {
+
+    }
+
+    @Override
+    public boolean waitUntilStarted(Duration timeout) {
+        return false;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof InMemorySubscription)) return false;
+        InMemorySubscription that = (InMemorySubscription) o;
+        return Objects.equals(id, that.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
+    }
+
+    @Override
+    public String toString() {
+        return new StringJoiner(", ", InMemorySubscription.class.getSimpleName() + "[", "]")
+                .add("id='" + id + "'")
+                .toString();
+    }
+}